flask/requirements.txt
torchvision
opencv-python
pandas==1.1.4
pillow==8.4.0
numpy==1.19.5
librosa==0.9.1
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
uWSGI==2.0.19.1
Werkzeug==1.0.1
